+++ src/ezdxf/addons/__init__.py
@@ -7,3 +7,4 @@ from .menger_sponge import MengerSponge
 from .sierpinski_pyramid import SierpinskyPyramid
 from .dimlines import LinearDimension, AngularDimension, ArcDimension, RadialDimension, dimstyles
 from .importer import Importer
+from .r12writer import r12writer